add search functionality

+++ lib/model/movie.dart
@@ -19,7 +19,7 @@ class Movie {
 
   Movie.fromJson(Map<String, dynamic> json)
       : id = json["id"],
-        popularity = json["popularity"],
+        popularity = json["popularity"].toDouble(),
         title = json["title"],
         backPoster = json["backdrop_path"],
         poster = json["poster_path"],
